feat(integration): logger out tps per 5s

+++ integration_tests/test_token.go
@@ -20,7 +20,7 @@ import (
 // TokenTest manage circulation of token
 type TokenTest struct {
 	accCnt   int
-	txCnt    int
+	txCnt    uint64
 	addrs    []string
 	accAddrs []string
 	quitCh   chan os.Signal
